Quick fix on order_item model

+++ app_pedidos/models/order_item.py
@@ -4,10 +4,9 @@ from .product import Product
 
 
 class OrderItem(models.Model):
-    #order = models.ForeignKey(Order, on_delete=models.CASCADE)
     order = models.ForeignKey(Order, on_delete=models.CASCADE, related_name='items', null=True, blank=True)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    price = models.BigIntegerField(default=0)
+    price = models.FloatField(default=0)
     quantityProduct = models.BigIntegerField(default=1)
 
     def __str__(self):
